+++ .idea/workspace.xml
@@ -6,10 +6,10 @@
   <component name="ChangeListManager">
     <list default="true" id="ee6614ce-2a79-474c-bc4c-ef3e0acb5ae0" name="Changes" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/frontend/lib/face_detection/guidline.dart" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/lib/face_detection/guidline.dart" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/frontend/lib/face_detection/face_result.dart" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/lib/face_detection/face_result.dart" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/frontend/lib/loading/loading_page0.dart" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/lib/loading/loading_page0.dart" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/frontend/lib/main.dart" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/lib/main.dart" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/frontend/lib/question/question3.dart" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/lib/question/question3.dart" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/frontend/lib/routines/routine_sucessfuly_create.dart" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/lib/routines/routine_sucessfuly_create.dart" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/frontend/lib/question/bsti.dart" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
